src/requirements.txt
-i https://pypi.org/simple
beautifulsoup4==4.8.1
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
feedparser==5.2.1
future==0.18.1
idna==2.8
psycopg2-binary==2.8.4
pycparser==2.19
python-telegram-bot==12.2.0
requests==2.22.0
six==1.12.0
soupsieve==1.9.4
sqlalchemy==1.3.10
tornado==6.0.3
urllib3==1.25.6
